requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1

# Machine Learning
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# Geolocation
geopy==2.4.1

# Development
pytest==8.3.3
httpx==0.27.2
